requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
cssselect2==0.7.0
Django==4.2.11
django-appconf==1.0.6
django-ckeditor==6.7.1
django-classy-tags==4.1.0
django-entangled==0.5.4
django-filer==3.1.1
django-formtools==2.5.1
django-fsm==2.8.1
django-js-asset==2.2.0
django-jazzmin
django-modeltranslation==0.18.11
django-parler==2.3
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-select2==8.1.2
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-alias==2.0.0
djangocms-attributes-field==3.0.0
djangocms-frontend==1.2.2
djangocms-text-ckeditor==5.1.5
djangocms-versioning==2.0.0
easy-thumbnails==2.8.5
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.0.2
html5lib==1.1
idna==3.6
lxml==5.1.0
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
StrEnum==0.4.15
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.10.0
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.6.0
pytelegrambotapi